Mahoji - Quest
Migrating questing command to a slash command.
